Xpetra: split MatrixFactory into two files
This works around an issue where Xpetra's library was missing
symbols for MatrixFactory2.  This occurs with the Intel OneAPI
compiler 2024.2.0.
