+++ packages/xpetra/src/Utils/ExplicitInstantiation/ETI_SC_LO_GO_NO_classes.cmake
@@ -6,6 +6,7 @@ APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::IO )
 APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::Matrix )
 APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::MatrixUtils )
 APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::MatrixFactory )
+APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::MatrixFactory2 )
 APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::MatrixMatrix )
 APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::ThyraUtils-.?if.defined[HAVE_XPETRA_THYRA] )
 APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::TpetraBlockCrsMatrix )
